requirements.txt
# SourcingBot Lite - Dependencies
# Core
google-genai>=1.0.0
pydantic>=2.0
httpx>=0.27
python-dotenv>=1.0

# API Server
fastapi>=0.115
uvicorn>=0.30

# Utilities
simpleeval>=1.0
